pyproject.toml
[tool.poetry]
name = "correction"
version = "2.0.0"
description = "Correction library for ALeRCE pipeline."
authors = ["ALeRCE Broker <alercebroker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
numpy = "~1.24.2"
pandas = "~1.5.3"
fastavro = { version = "~1.4.12", optional = true }
prometheus-client = { version = "~0.16.0", optional = true }
confluent-kafka = { version = "~2.0.2", optional = true }
apf_base = { version = "~2.4.0", optional = true }
pyroscope-io = "^0.8.4"

[tool.poetry.extras]
apf = ["fastavro", "prometheus-client", "confluent-kafka", "apf_base"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-docker = "^1.0.1"
black = "^22.12.0"

[tool.poetry.scripts]
run-step = { callable = "correction._step:run_step", extras = ["apf"] }

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
